pyproject.toml
[tool.poetry]
name = "gdb-sync"
version = "0.1.0"
description = ""
authors = ["Abraham Yusuf <abraham.yusuf@ehealthnigeria.org>"]

[tool.poetry.dependencies]
python = "^3.7"
psycopg2-binary = "^2.7"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
pylint = "^2.1"
black = "^18.3-alpha.0"

[tool.poetry.scripts]
gdb-sync = "gdb_sync:__main__.main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
